% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning.R
\name{impBorutaCML}
\alias{impBorutaCML}
\title{Feature importance from Consensus Machine Learning (CML) using the Boruta method}
\usage{
impBorutaCML(
  x = x,
  y = y,
  seed,
  algo.opt = c("lasso", "rf", "svm", "xgb"),
  ranksummary = "median"
)
}
\arguments{
\item{x}{Valid data object.}

\item{y}{Vector of classes among instances in data object.}

\item{seed}{Random seed to use in model fitting.}

\item{algo.opt}{Algorithms to use for consensus method.}

\item{ranksummary}{Method to compute rank summary among algorithms used, currently supports 'median'.}
}
\value{
Consensus feature importance.
}
\description{
Returns feature importance from fitting a model using CML with the Boruta method.
}
