% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning.R
\name{impCML}
\alias{impCML}
\title{Feature importance from Consensus Machine Learning (CML)}
\usage{
impCML(
  df,
  classes,
  trainindices,
  seed,
  algo.opt = c("lasso", "rf", "svm", "xgb"),
  standtable = FALSE,
  ranksummary = "median"
)
}
\arguments{
\item{df}{Valid data object.}

\item{classes}{Vector of class specifying test or training set among instances.}

\item{trainindices}{Indices of traning instances in data.}

\item{seed}{Random seed to use in model fitting.}

\item{algo.opt}{Algorithms to use for consensus method.}

\item{standtable}{Boolean, whether to return a standard output table summarizing results.}

\item{ranksummary}{Method to compute rank summary among algorithms used, currently supports 'median'.}
}
\value{
Consensus feature importance.
}
\description{
Returns feature importance from fitting a model using CML
}
