% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning.R
\name{impLasso}
\alias{impLasso}
\title{Feature importance from lasso}
\usage{
impLasso(df, classes, trainindices, seed = 2019)
}
\arguments{
\item{df}{Valid data object.}

\item{classes}{Vector of class specifying test or training set among instances.}

\item{trainindices}{Which instances correspond to the training set.}

\item{seed}{Random seed to use in model fitting.}
}
\value{
Lasso feature importances.
}
\description{
Returns feature importance from fitting a model using lasso.
}
