% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning.R
\name{impRF}
\alias{impRF}
\title{Feature importance from random forest}
\usage{
impRF(df, classes, ntrees = 100, seed = 2019)
}
\arguments{
\item{df}{Valid data object.}

\item{classes}{Vector of class specifying test or training set among instances.}

\item{ntrees}{Number of trees to use in random forest run.}

\item{seed}{Random seed to use in model fitting.}
}
\value{
Random forest feature importances.
}
\description{
Returns feature importance from fitting a model using random forests.
}
