% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning.R
\name{impSVM}
\alias{impSVM}
\title{Feature importance from Support Vector Machines (SVM)}
\usage{
impSVM(df, classes, seed = 2019)
}
\arguments{
\item{df}{Valid data object.}

\item{classes}{Vector of class specifying test or training set among instances.}

\item{seed}{Random seed to use in model fitting.}
}
\value{
SVM feature importances.
}
\description{
Returns feature importance from fitting a model using SVM
}
