% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning.R
\name{runLasso}
\alias{runLasso}
\title{Run lasso}
\usage{
runLasso(seset, seed = 2019)
}
\arguments{
\item{seset}{A SummarizedExperiment object.}
}
\value{
A `resultslist` object containing the fitted model and evaluation info.
}
\description{
This function runs lasso using a SummarizedExperiment object. Credit base code: Jenny Smith.
}
