% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning.R
\name{runSVM}
\alias{runSVM}
\title{Fit a SupportVectorMachines model.}
\usage{
runSVM(seed, kerneltype = "linear", seset, weightfilt = FALSE)
}
\arguments{
\item{seed}{Set a random seed for the SVM run.}

\item{kerneltype}{Choose the kernel type for the SVM run.}

\item{weightfilt}{False or numeric float. Float specifies the fraction of absolute weights to retain in the fitted model.}

\item{sest}{A valid SummarizedExperiment object, including variable for test and training subsets.}
}
\value{
A `resultslist` object containing the fitted model, predictions, and performacne metrics.
}
\description{
This function runs lasso using a SummarizedExperiment object. Credit base code: Jenny Smith.
}
